refactor organization: Use of lib.rs file to declare root modules
