+++ src/base/commands/commands/jumps.rs
@@ -1,7 +1,7 @@
 use crate::base::commands::CommandTrait;
 use crate::base::states::states::{self, State};
-use crate::commands::{Command, Commands};
-use crate::App;
+use crate::base::commands::{Command, Commands};
+use crate::app::App;
 use std::sync::Arc;
 
 impl Commands {
