+++ src/lib.rs
@@ -1,0 +1,7 @@
+pub mod app;
+pub mod base;
+pub mod config;
+pub mod input;
+pub mod logger;
+pub mod utils;
+pub mod view;
