+++ tests/integration/mod.rs
@@ -1,8 +1,12 @@
+
 #[cfg(test)]
 mod file_handler {
+    use treq::base::os::handler::FileHandler;
+
     #[test]
     #[ignore]
     fn sgodasdasdsa() {
+        let _empty = FileHandler::default();
         assert_eq!(4, 4)
     }
 }
